pyproject.toml
[project]
name = "nobious-chatbot"
version = "0.1.0"
description = "Nobious IMS AI Chat Service"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0",
    "flask-limiter>=3.5",
    "langgraph>=0.2",
    "langchain>=0.2",
    "langchain-aws>=0.2",
    "bm25s>=0.2",
    "gitpython>=3.1",
    "pydantic-settings>=2.0",
    "click>=8.1",
    "httpx>=0.27",
    "PyStemmer>=2.2",
    "langchain-openai>=1.1.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
